#202 Infrastructure - Refactor Permission Repository to Permission Provider
